Jar Update #01 V2
Download now and play!

Requires Java 8 or higher to run.

Might be opened by double-clicking this jar file.
If desired to see the message logs it can be ran by opening the CMD/Terminal prompt in the respective folder of the jar and run the line: java -jar SnakeGame.jar

Enjoy it! :)